# Cache the block which most recently successfully allocated something