@robertosfield 2024-09-25
ooo that's a big improvement.  I think this change might need some refinement, but broadly it looks a sensible direction to follow.  Once I've wrapped up my present dev work I'll have do a proper review.