@AnyOldName3 2024-09-25
Generally, I'd be in favour of using an off-the-shelf allocator instead of a custom one so we could be relatively sure there weren't any undiscovered pathological cases that people would run into, but I did a review of what's available, and the ones that natively supported something like the VSG affinity system (which they mostly referred to as *multiple heaps*) either were tied to a specific platform or didn't support allocating and freeing from one heap on different threads. Mimalloc has https://github.com/microsoft/mimalloc/issues/925 where one of the maintainers states that they eventually intend to make it possible, so the situation may improve, but we're stuck with something custom in the short term.