@AnyOldName3 2024-09-26
The data from another set of runs (as the first were missing some lines due to Windows 10 still shipping PowerShell with PSReadLine 2.0 installed, which has a bug that was fixed in 2018 that makes some lines disappear from the scrollback buffer with some terminals), but formatted as a table so it's vaguely readable:

## Original
| Allocator | Pass       | Construction time | Traversal time | Destruction time | Total time | Construction rate | Traversal rate | Destruction rate |
| --------- | ---------- | ----------------- | -------------- | ---------------- | ---------- | ----------------- | -------------- | ---------------- |
| std       | 0          | 1.02466           | 3.0281         | 1.30101          | 5.35376    | 2.18E+07          | 7.39E+07       | 1.72E+07         |
| std       | 1          | 1.02495           | 3.06137        | 1.35196          | 5.43827    | 2.18E+07          | 7.31E+07       | 1.65E+07         |
| std       | 2          | 1.01582           | 3.05855        | 1.31805          | 5.39242    | 2.20E+07          | 7.31E+07       | 1.70E+07         |
| std       | Average    | 1.02181           | 3.04934        | 1.32367          | 5.39482    | 2.19E+07          | 7.34E+07       | 1.69E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| std       | 0          | 1.00797           | 3.02305        | 1.27342          | 5.30444    | 2.22E+07          | 7.40E+07       | 1.76E+07         |
| std       | 1          | 1.08211           | 3.10958        | 1.34273          | 5.53441    | 2.07E+07          | 7.19E+07       | 1.67E+07         |
| std       | 2          | 1.027             | 3.10732        | 1.29293          | 5.42725    | 2.18E+07          | 7.20E+07       | 1.73E+07         |
| std       | Average    | 1.03902           | 3.07998        | 1.30303          | 5.42203    | 2.15E+07          | 7.26E+07       | 1.72E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| std       | 0          | 1.00203           | 3.10911        | 1.32102          | 5.43215    | 2.23E+07          | 7.19E+07       | 1.69E+07         |
| std       | 1          | 1.04492           | 3.01074        | 1.29855          | 5.35421    | 2.14E+07          | 7.43E+07       | 1.72E+07         |
| std       | 2          | 1.04625           | 3.05348        | 1.35697          | 5.4567     | 2.14E+07          | 7.33E+07       | 1.65E+07         |
| std       | Average    | 1.03107           | 3.05777        | 1.32551          | 5.41435    | 2.17E+07          | 7.32E+07       | 1.69E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| intrusive | 0          | 0.959542          | 0.983801       | 1.04838          | 2.99173    | 2.33E+07          | 2.27E+08       | 2.13E+07         |
| intrusive | 1          | 76.78             | 1.03939        | 1.05049          | 78.8699    | 291347            | 2.15E+08       | 2.13E+07         |
| intrusive | 2          | 102.748           | 0.996635       | 1.0704           | 104.815    | 217714            | 2.24E+08       | 2.09E+07         |
| intrusive | Average    | 60.1625           | 1.00661        | 1.05643          | 62.2255    | 371820            | 2.22E+08       | 2.12E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| intrusive | 0          | 0.932389          | 0.97655        | 1.03791          | 2.94685    | 2.40E+07          | 2.29E+08       | 2.16E+07         |
| intrusive | 1          | 75.3725           | 1.0134         | 1.04399          | 77.4298    | 296788            | 2.21E+08       | 2.14E+07         |
| intrusive | 2          | 97.1932           | 1.00882        | 1.04227          | 99.2442    | 230156            | 2.22E+08       | 2.15E+07         |
| intrusive | Average    | 57.8327           | 0.999589       | 1.04139          | 59.8736    | 386799            | 2.24E+08       | 2.15E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| intrusive | 0          | 0.932609          | 0.990708       | 1.0673           | 2.99061    | 2.40E+07          | 2.26E+08       | 2.10E+07         |
| intrusive | 1          | 79.665            | 1.03739        | 1.06861          | 81.771     | 280796            | 2.16E+08       | 2.09E+07         |
| intrusive | 2          | 98.5477           | 0.97443        | 1.06761          | 100.59     | 226993            | 2.30E+08       | 2.10E+07         |
| intrusive | Average    | 59.7151           | 1.00084        | 1.06784          | 61.7838    | 374606            | 2.24E+08       | 2.09E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| std       | Average 0  | 1.011553333       | 3.05342        | 1.298483333      | 5.36345    | 22116133.33       | 73273033.33    | 17231433.33      |
|           | Average 12 | 1.040175          | 3.06684        | 1.326865         | 5.4338767  | 21514816.67       | 72949250       | 16865133.33      |
|           |            |                   |                |                  |            |                   |                |                  |
| intrusive | Average 0  | 0.941513333       | 0.983686333    | 1.051196667      | 2.9763967  | 23763533.33       | 227414000      | 21282933.33      |
|           | Average 12 | 88.3844           | 1.0116775      | 1.057228333      | 90.453317  | 257299            | 221225000      | 21161466.67      |

## New
| Allocator | Pass       | Construction time | Traversal time | Destruction time | Total time | Construction rate | Traversal rate | Destruction rate |
| --------- | ---------- | ----------------- | -------------- | ---------------- | ---------- | ----------------- | -------------- | ---------------- |
| std       | 0          | 0.988737          | 2.97296        | 1.28652          | 5.24821    | 2.26E+07          | 7.52E+07       | 1.74E+07         |
| std       | 1          | 1.05222           | 2.93875        | 1.29863          | 5.28959    | 2.13E+07          | 7.61E+07       | 1.72E+07         |
| std       | 2          | 1.0644            | 2.97248        | 1.2872           | 5.32409    | 2.10E+07          | 7.53E+07       | 1.74E+07         |
| std       | Average    | 1.03512           | 2.9614         | 1.29078          | 5.28729    | 2.16E+07          | 7.55E+07       | 1.73E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| std       | 0          | 1.03657           | 2.95491        | 1.28253          | 5.27401    | 2.16E+07          | 7.57E+07       | 1.74E+07         |
| std       | 1          | 1.06806           | 2.95792        | 1.28121          | 5.30719    | 2.09E+07          | 7.56E+07       | 1.75E+07         |
| std       | 2          | 1.03851           | 3.00261        | 1.34675          | 5.38787    | 2.15E+07          | 7.45E+07       | 1.66E+07         |
| std       | Average    | 1.04772           | 2.97181        | 1.3035           | 5.32303    | 2.14E+07          | 7.53E+07       | 1.72E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| std       | 0          | 1.03263           | 2.97959        | 1.32292          | 5.33515    | 2.17E+07          | 7.51E+07       | 1.69E+07         |
| std       | 1          | 1.01171           | 2.94223        | 1.29528          | 5.24921    | 2.21E+07          | 7.60E+07       | 1.73E+07         |
| std       | 2          | 1.01288           | 2.97323        | 1.30579          | 5.2919     | 2.21E+07          | 7.52E+07       | 1.71E+07         |
| std       | Average    | 1.01907           | 2.96501        | 1.308            | 5.29208    | 2.20E+07          | 7.54E+07       | 1.71E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| intrusive | 0          | 0.918469          | 0.947844       | 1.09361          | 2.95993    | 2.44E+07          | 2.36E+08       | 2.05E+07         |
| intrusive | 1          | 0.729437          | 0.946425       | 1.08236          | 2.75822    | 3.07E+07          | 2.36E+08       | 2.07E+07         |
| intrusive | 2          | 0.74219           | 0.907451       | 1.0768           | 2.72644    | 3.01E+07          | 2.47E+08       | 2.08E+07         |
| intrusive | Average    | 0.796698          | 0.933907       | 1.08426          | 2.81486    | 2.81E+07          | 2.40E+08       | 2.06E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| intrusive | 0          | 0.928767          | 0.937689       | 1.05957          | 2.92602    | 2.41E+07          | 2.39E+08       | 2.11E+07         |
| intrusive | 1          | 0.720648          | 0.979823       | 1.08049          | 2.78096    | 3.10E+07          | 2.28E+08       | 2.07E+07         |
| intrusive | 2          | 0.735943          | 0.924255       | 1.06408          | 2.72428    | 3.04E+07          | 2.42E+08       | 2.10E+07         |
| intrusive | Average    | 0.79512           | 0.947256       | 1.06805          | 2.81042    | 2.81E+07          | 2.36E+08       | 2.09E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| intrusive | 0          | 0.920807          | 0.931522       | 1.06723          | 2.91955    | 2.43E+07          | 2.40E+08       | 2.10E+07         |
| intrusive | 1          | 0.72225           | 0.965551       | 1.06364          | 2.75144    | 3.10E+07          | 2.32E+08       | 2.10E+07         |
| intrusive | 2          | 0.734446          | 0.912303       | 1.06888          | 2.71563    | 3.05E+07          | 2.45E+08       | 2.09E+07         |
| intrusive | Average    | 0.792501          | 0.936458       | 1.06658          | 2.79554    | 2.82E+07          | 2.39E+08       | 2.10E+07         |
|           |            |                   |                |                  |            |                   |                |                  |
| std       | Average 0  | 1.019312333       | 2.969153333    | 1.297323333      | 5.28579    | 21955833.33       | 75341100       | 17246200         |
|           | Average 12 | 1.041296667       | 2.964536667    | 1.302476667      | 5.3083083  | 21492633.33       | 75461416.67    | 17179183.33      |
|           |            |                   |                |                  |            |                   |                |                  |
| intrusive | Average 0  | 0.922681          | 0.939018333    | 1.07347          | 2.9351667  | 24244700          | 238235666.7    | 20842433.33      |
|           | Average 12 | 0.730819          | 0.939301333    | 1.072708333      | 2.7428283  | 30612283.33       | 238346333.3    | 20854450         |

In light of this, I can say that:
* Reuse of blocks has changed from ludicrously slow to being faster than using fresh blocks.
* There's a possibility that it's slightly improved traversal times when blocks are reused, but it's not a significant enough effect to be conclusive.
* It doesn't look like anything was made worse.

The OS doing secret stuff under the hood should have been prevented by running the std allocator test first as that *should* have meant there was enough memory that wasn't being used for disk cache or anything else that might need paging out.